move search result out as own component - using props and store to connect component and app state
