+++ UI/src/components/SearchSpotify.vue
@@ -87,98 +87,11 @@
     </div>
     <hr class="divide-solid my-2">
     <!-- search results -->
-    <div class="overflow-y-auto overflow-x-hidden" style="height: calc(100vh - 163px);">
-      <template v-if="activeSearchView !== 'new-releases'">
-        <div
-          v-if="activeSearchArr.length !== 0"
-          class="grid grid-cols-3 gap-4 px-4 pt-1"
-        >
-          <template v-for="(item ,i) in activeSearchArr">
-            <div v-if="item !== null" :key="i" :id="item.name ? item.name : ''">
-              <!-- artists -->
-              <div
-                v-if="activeSearchView === 'artists'"
-                class="box rounded h-32 w-full relative"
-                :class="item.images.length > 0 ? 'hover:opacity-50 cursor-pointer' : 'bg-green-700'"
-              >
-                <img
-                  v-if="item.images.length > 0"
-                  :id="item.id"
-                  :src="item.images[1].url"
-                  :alt="item.name"
-                  loading=lazy
-                  class="object-cover w-full h-full rounded"
-                  @click="addToCanvas(item.images[0].url), addToSelection(item.name, item.images[0].url, item.id)"
-                  v-longclick="() => setLongpress()"
-                >
-                <Icons v-else name="artist"/>
-              </div>
-              <!-- tracks -->
-              <div
-                v-else-if="activeSearchView === 'tracks'"
-                class="box rounded h-32 w-full relative"
-                :class="item.album.images.length > 0 ? 'hover:opacity-50 cursor-pointer' : 'bg-green-700'"
-              >
-                  <img
-                    v-if="item.album.images.length > 0"
-                    :id="item.id"
-                    :src="item.album.images[1].url"
-                    :alt="item.name"
-                    loading=lazy
-                    class="object-cover w-full h-full rounded "
-                    @click="handleClick(item.album.images[0].url, item.preview_url), addToSelection(item.name, item.album.images[0].url, item.id)"
-                    v-longclick="() => setLongpress()"
-                  >
-                  <Icons v-else name="track"/>
-              </div>
-              <div class="text-xs pt-1 font-semibold">{{ item.name }}</div>
-              <p class="text-xs">
-                <template v-for="(artist, i) in item.artists" class="text-xs">{{ i > 0 ? `, ${artist.name}` : artist.name }}</template>
-              </p>
-            </div>
-          </template>
-        </div>
-        <div v-else class="flex justify-center items-center h-full">
-          <div 
-            class="text-center"
-            style="max-width:240px"
-          >
-            <template v-if="activeSearchView === 'artists'">
-              <h1 class="text-lg pb-2 font-semibold">Search Artists</h1>
-              <p class="text-xs">Get Spotify Catalog information about artists that match a keyword string.</p>
-            </template>
-            <template v-if="activeSearchView === 'tracks'">
-              <h1 class="text-lg pb-2 font-semibold">Search Tracks</h1>
-              <p class="text-xs pb-1">Get Spotify Catalog information about tracks that match a keyword string and play track previews.</p>
-            </template>
-          </div>
-        </div>
-      </template>
-      <template v-if="activeSearchView === 'new-releases'">
-        <div class="grid grid-cols-3 gap-4 px-4 pt-1">
-          <!-- new releases -->
-          <div v-for="(item ,i) in newReleasesArr" :key="i" :id="item.name">
-            <div class="box rounded h-32 w-full relative hover:opacity-50 cursor-pointer">
-              <img
-                v-if="item.images.length > 0"
-                :id="item.id"
-                :src="item.images[1].url"
-                :alt="item.name"
-                loading=lazy
-                class="object-cover w-full h-full rounded"
-                @click="addToCanvas(item.images[0].url), addToSelection(item.name, item.images[0].url, item.id)"
-                v-longclick="() => setLongpress()"
-              >
-              <Icons v-else name="artist"/>
-            </div>
-            <div class="text-xs pt-1 pb-1 font-semibold">{{ item.name }}</div>
-            <p class="text-xs">
-              <template v-for="(artist, i) in item.artists" class="text-xs">{{ i > 0 ? `, ${artist.name}` : artist.name }}</template>
-            </p>
-          </div>
-        </div>
-      </template>
-    </div>
+    <SearchResults
+      :active_search_array="activeSearchArr"
+      :new_releases_array="newReleasesArr"
+      :active_search_view="activeSearchView"
+    ></SearchResults>
     <!-- controls -->
     <div class="absolute bottom-0 left-0 right-0 bg-white">
       <hr class="divide-solid my-2">
@@ -244,7 +157,9 @@
 
 <script>
 import axios from 'axios';
-import { notify, createImage, createImageArray } from "../helpers/figma-messages";
+import { mapState } from 'vuex'
+import { notify, createImageArray } from "../helpers/figma-messages";
+import SearchResults from "@/components/SearchResults";
 import Icons from "./Icons.vue";
 import Menu from "./Menu.vue"
 
@@ -255,7 +170,11 @@ var data = qs.stringify({
 
 export default {
   name: "SearchSpotify",
-  components: {Icons, Menu},
+  components: {
+    Icons,
+    Menu,
+    SearchResults
+  },
   data() {
     return {
       activeSearchView: 'new-releases',
@@ -272,11 +191,6 @@ export default {
         width: 800,
         height: 800
       },
-      arraySel: [],
-      longpress: false,
-      playing: false,
-      clickCounter: 0,
-      timer: null,
       tooltip: false
     };
   },
@@ -294,6 +208,11 @@ export default {
       this.resetLongpress()
     }
   },
+  computed: mapState({
+    arraySel: 'arraySel',
+    longpress: 'longpress',
+    playing: 'playing'
+  }),
   methods: {
     enterSearch(e) {
       if (e.code === 'Enter') {
@@ -302,7 +221,7 @@ export default {
       }
     },
     getNewReleases() {
-      this.arraySel = []
+      this.$store.commit('arraySel', [])
       var auth = process.env.VUE_APP_SPOTIFY_AUTH_BASIC
       axios({
         method: 'post',
@@ -332,7 +251,7 @@ export default {
       });
     },
     querySpotify() {
-      this.arraySel = []
+      this.$store.commit('arraySel', [])
       if (this.searchQuery.length !== 0) {
         var auth = process.env.VUE_APP_SPOTIFY_AUTH_BASIC
         axios({
@@ -420,66 +339,6 @@ export default {
       var input = document.querySelector('input')
       input.focus()
     },
-    handleClick(imageURL, trackURL) {
-      this.clickCounter++
-      if (this.clickCounter === 1) {
-        this.timer = setTimeout(() => {
-          this.clickCounter = 0
-          this.addToCanvas(imageURL)
-        }, 200);
-      } else if (this.clickCounter === 2) {
-        clearTimeout(this.timer)
-        this.clickCounter = 0
-        if (trackURL) {
-          this.playTrack(trackURL)
-        } else {
-          notify(`Track not available`, {timeout: 1500})
-        }
-      }
-    },
-    addToCanvas(imageURL) {
-      if (this.longpress === false) {
-        axios({
-          method: 'get',
-          url: imageURL,
-          responseType: 'arraybuffer'
-        })
-        .then( response => {
-          var arrayBufferView = new Uint8Array( response.data );
-          // send data to figma
-          createImage(arrayBufferView, this.size)
-        })
-      }
-    },
-    addToSelection(name, url, id) {
-      if (this.longpress === true) {
-        var obj = {
-          id: id,
-          name: name,
-          url: url
-        }
-        // check for duplicate
-        var dupObj = this.arraySel.find(obj => {
-          return obj.id === id
-        })
-
-        if (dupObj) {
-          // if already exists then remove from array
-          var index = this.arraySel.indexOf(dupObj)
-          this.arraySel.splice(index, 1)
-          var clickObj = document.getElementById(id)
-          clickObj.classList.remove('opacity-30')
-          clickObj.parentElement.classList.remove('selected-blue')
-        }
-        else {
-          // else add into array
-          var select = document.getElementById(id)
-          select.classList.add('opacity-30')
-          select.parentElement.classList.add('selected-blue')
-          this.arraySel.push(obj)
-        }        
-      }
-    },
     addSelectionToCanvas() {
       var imgBufferArray = []
       var promises = []
@@ -503,11 +362,8 @@ export default {
         })
       }
     },
-    setLongpress() {
-      this.longpress = true;
-    },
     resetLongpress() {
-      this.longpress = false
+      this.$store.commit('longpress', false)
       // remove all opacity classes
       var borderArr = document.querySelectorAll('.selected-blue')
       borderArr.forEach(el => {
@@ -517,31 +373,12 @@ export default {
       opacityArr.forEach(el => {
         el.classList.remove('opacity-30')
       })
-      this.arraySel = []
-    },
-    playTrack(url) {
-      clearTimeout(this.timer)
-      var audio = document.getElementById('audioPlayer')
-      audio.src = url
-      if (audio.src) {
-        var playPromise = audio.play()
-        if (playPromise !== undefined) {
-          playPromise.then(() => {
-            this.playing = true
-          }).catch((err) => {
-            this.playing = false
-            console.log(err)
-          })
-        }
-        audio.onended = () => {
-          this.playing = false
-        }
-      }
+      this.$store.commit('arraySel', [])
     },
     stopTrack() {
       var audio = document.getElementById('audioPlayer')
       audio.src = ''
-      this.playing = false
+      this.$store.commit('playing', false)
     }
   }
 };
